+++ src/widgets/font-detail/ui/FontDetailContent.jsx
